% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkinp.R
\name{chkinp}
\alias{chkinp}
\title{Check input data for IPI}
\usage{
chkinp(stations, phab, qa = TRUE, allerr = TRUE)
}
\arguments{
\item{stations}{\code{data.frame} of input station data}

\item{phab}{\code{data.frame} of input physical habitat data}

\item{qa}{logical value passed from \code{\link{IPI}} to suppress error checks on relevant metrics}

\item{allerr}{logical indicating if all errors are returned or the first encountered}
}
\value{
An error message is returned if the input data are not correctly formatted. If a dataset has multiple errors, only the first is returned.
}
\description{
Check input data for IPI
}
\details{
This function checks the following:
\describe{
\item{}{No duplicate station rows in \code{\link{stations}}}
\item{}{All stations in \code{\link{stations}} are in \code{\link{phab}} and the converse}
\item{}{All required fields in \code{\link{stations}} and \code{\link{phab}}}
\item{}{All required PHAB variables are present in the \code{variable} field of \code{\link{phab}} for each station and sample date.  An exception is made for \code{XC}, \code{PCT_POOL}, and \code{XFC_ALG}, which are not necessary for the IPI but are used for optional quality assurance checks.}
\item{}{No duplicate results for PHAB variables at each station and sample date}
\item{}{All input variables for \code{\link{stations}} and \code{\link{phab}} are non-negative, excluding elevation variables in \code{\link{stations}} which may be negative if below sea level (i.e., some locations in southeast California). The variables \code{XBKF_W} and \code{Ev_FlowHab} in \code{\link{phab}} must also be greater than zero.}
}
}
\examples{
chkinp(stations, phab)
}
