% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_AqHab.R
\docType{data}
\name{H_AqHab}
\alias{H_AqHab}
\title{Random forest model for H_AqHab}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
H_AqHab
}
\description{
Random forest model for H_AqHab
}
\examples{
data(H_AqHab)
}
\keyword{datasets}
