% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPI.R
\name{IPI}
\alias{IPI}
\title{IPI scores}
\usage{
IPI(stations, phab)
}
\arguments{
\item{stations}{\code{data.frame} of input station data}

\item{phab}{\code{data.frame} of input physical habitat data}
}
\value{
results
}
\description{
IPI scores
}
\examples{
IPI(stations, phab)
}
