% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCT_SAFN.R
\docType{data}
\name{PCT_SAFN}
\alias{PCT_SAFN}
\title{Random forest model for percent sands and fines}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
PCT_SAFN
}
\description{
Random forest model for percent sands and fines
}
\examples{
data(PCT_SAFN)
}
\keyword{datasets}
