% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phab.R
\docType{data}
\name{phab}
\alias{phab}
\title{Example physical habitat data}
\format{A \code{\link[base]{data.frame}} with 472 ows and 19 variables.}
\usage{
phab
}
\description{
Example physical habitat data
}
\examples{
data(phab)
}
\keyword{datasets}
