% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcal.R
\docType{data}
\name{refcal}
\alias{refcal}
\title{Calibration data}
\format{A \code{\link{data.frame}} of PHAB variables at reference and stressed sites used to estimate scoring limits
\describe{
  \item{\code{Variable}}{chr} name of the PHAB metric
  \item{\code{StationCode}}{chr} unique station identifier 
  \item{\code{SampleID2}}{chr} unique identifier of the sampling event
  \item{\code{SiteSet}}{chr} indicating reference or stressed site
  \item{\code{Result}}{num} resulting metric value
  \item{\code{Predicted}}{num} predicted metric value
}}
\usage{
refcal
}
\description{
Calibration data used to establish scoring limits for metrics
}
\examples{
data(refcal)
}
\keyword{datasets}
