% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Example station data}
\format{A \code{\link[base]{data.frame}} with 6 rows and 112 variables.}
\usage{
stations
}
\description{
Example station data
}
\examples{
data(stations)
}
\keyword{datasets}
