% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMG.R
\docType{data}
\name{XCMG}
\alias{XCMG}
\title{Random forest model for XCMG}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
XCMG
}
\description{
Random forest model for XCMG
}
\examples{
data(XCMG)
}
\keyword{datasets}
