% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkinp.R
\name{chkinp}
\alias{chkinp}
\title{Check input data for IPI}
\usage{
chkinp(station, phab)
}
\arguments{
\item{phab}{\code{data.frame} of input physical habitat data}

\item{stations}{\code{data.frame} of input station data}
}
\value{
An error message is returned if the input data are not correctly formatted. If a dataset has multiple errors, only the first is returned.
}
\description{
Check input data for IPI
}
\details{
This function checks the following:
\describe{
\item{}{No duplicate station rows in \code{\link{stations}}}
\item{}{All stations in \code{\link{stations}} are in \code{\link{phab}} and the converse}
\item{}{All required fields in \code{\link{stations}} and \code{\link{phab}}}
\item{}{All required PHAB variables are present in the \code{variable} field of \code{\link{phab}} for each station and sample date}
\item{}{No duplicate results for PHAB variables at each station and sample date}
}
}
\examples{
chkinp(stations, phab)
}
