% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPI.R
\name{IPI}
\alias{IPI}
\title{IPI scores}
\usage{
IPI(stations, phab, qa = TRUE)
}
\arguments{
\item{stations}{\code{data.frame} of input station data}

\item{phab}{\code{data.frame} of input physical habitat data}

\item{qa}{logical indicating of quality assurance columns are appended to the output}
}
\value{
A results \code{data.frame} of IPI and metric scores, quality assurance results are returned if \code{qa = TRUE} (default)
}
\description{
Estimate IPI scores for PHAB using station and physical habitat data
}
\examples{
IPI(stations, phab)
}
