% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPI.R
\name{IPI}
\alias{IPI}
\title{IPI scores}
\usage{
IPI(stations, phab, qa = TRUE, allerr = TRUE, log = FALSE)
}
\arguments{
\item{stations}{\code{data.frame} of input station data}

\item{phab}{\code{data.frame} of input physical habitat data}

\item{qa}{logical indicating of quality assurance columns are appended to the output}

\item{allerr}{logical indicating if all errors are returned or the first encountered}

\item{log}{logical indicating if errors are printed to log in the workign directory, applies only if \code{allerr = TRUE}}
}
\value{
A results \code{data.frame} of IPI and metric scores, quality assurance results are returned if \code{qa = TRUE} (default)
}
\description{
Estimate IPI scores for PHAB using station and physical habitat data
}
\examples{
IPI(stations, phab)
}
