% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phab.R
\docType{data}
\name{phab}
\alias{phab}
\title{Example physical habitat data}
\format{A \code{\link[base]{data.frame}} with 472 rows and 5 variables. Required variables for the \code{\link{IPI}} function are \code{StationCode}, \code{SampleDate}, \code{Variable}, \code{Result}, and \code{Count_Calc}.  Required variables in the \code{Variable} field are \code{XSLOPE}, \code{XBKF_W}, \code{H_AqHab}, \code{PCT_SAFN}, \code{XCMG}, \code{Ev_FlowHab}, \code{H_SubNat}, \code{XC}, \code{PCT_POOL}, \code{XFC_ALG}, and \code{PCT_RC}.}
\usage{
phab
}
\description{
Example physical habitat data
}
\examples{
data(phab)
}
\keyword{datasets}
