% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Example station data}
\format{A \code{\link[base]{data.frame}} with 6 rows and 11 variables.  Required variables for the \code{\link{IPI}} function are \code{StationCode}, \code{MAX_ELEV}, \code{AREA_SQKM}, \code{ELEV_RANGE}, \code{MEANP_WS}, \code{New_Long}, \code{SITE_ELEV}, \code{KFCT_AVE}, \code{New_Lat}, \code{MINP_WS}, and \code{PPT_00_09}.}
\usage{
stations
}
\description{
Example station data
}
\examples{
data(stations)
}
\keyword{datasets}
